fix(reasonix): bound session detail metadata reads
